LECTURE_NOTES/MSTPL/first_year_master/2025-2026/open_word_project/owproject/pyproject.toml
[project]
name = "owproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "torch>=2.10.0",
    "transformers>=5.0.0",
]
